requirements.txt
numpy
torch
torchvision
absl_py
accelerate
gradio_imageslider
gradio==4.29.0
gin_config
imageio
imageio[ffmpeg]
matplotlib
mediapy
opencv_contrib_python
opencv_python
Pillow
trimesh
transformers
pymeshlab
xatlas
plyfile
rawpy
ninja
scipy
scikit-image
scikit-learn
tqdm
wandb